added the possibility to mask description + css improvement for the fullscreen button
